pyproject.toml
[tool.ruff]
line-length = 100

[tool.poetry]
name = "studyproject"
version = "0.1.0"
description = ""
authors = ["Artem <Gakerchannel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.115.6"}
uvicorn = "^0.32.1"
python-multipart = "^0.0.19"
pydantic-settings = "^2.6.1"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
black = "^24.10.0"
pyjwt = "^2.10.1"
bcrypt = "4.0.1"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^2.10.3"}
redis = "^5.2.1"
fastapi-cache2 = "^0.2.2"
celery = "^5.4.0"
pillow = "^11.0.0"
aiosmtplib = "^3.0.2"
pytest-asyncio = "^0.25.2"
python-dotenv = "^1.0.1"
pytest = "^8.3.4"
pytest-dotenv = "^0.5.2"
pytest-order = "^1.3.0"
ruff = "^0.9.2"
validate-email-address = "^1"
py3dns = "^4.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
